[ETCM-792] Extract monitoring http endpoints to trait
